Update int sizes.

+++ main.cpp
@@ -104,7 +104,7 @@ void initWorld(uint8_t worldIndex) {
 
     // Populate the world data ..
     
-    for (uint8_t i = 0; i < Constants::mapTileWidth * Constants::mapTileHeight / 2; i++) {
+    for (uint16_t i = 0; i < Constants::mapTileWidth * Constants::mapTileHeight / 2; i++) {
         
         worldMap[i] = Data::worldMaps[worldIndex][i];
     }
